backend/requirements.txt
fastapi
uvicorn
transformers
torch
llama-cpp-python
pydantic
